pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-session>=0.8.0",
    "google-genai>=1.39.1",
    "werkzeug>=3.1.3",
    "SpeechRecognition>=3.10.4",
    "pydub>=0.25.1",
]
